feat: Provided registration of points upon completing an action and fixed various issues in scripts to handle update of actions

+++ og_db/migrations/10_create_players.up.sql
@@ -43,11 +43,11 @@ CREATE TABLE players_technologies (
 -- Create a table representing the points accumulated for a player.
 CREATE TABLE players_points (
   player uuid NOT NULL,
-  economy_points integer NOT NULL DEFAULT 0,
-  research_points integer NOT NULL DEFAULT 0,
-  military_points_built integer NOT NULL DEFAULT 0,
-  military_points_lost integer NOT NULL DEFAULT 0,
-  military_points_destroyed integer NOT NULL DEFAULT 0,
+  economy_points numeric(15, 5) NOT NULL DEFAULT 0,
+  research_points numeric(15, 5) NOT NULL DEFAULT 0,
+  military_points_built numeric(15, 5) NOT NULL DEFAULT 0,
+  military_points_lost numeric(15, 5) NOT NULL DEFAULT 0,
+  military_points_destroyed numeric(15, 5) NOT NULL DEFAULT 0,
   FOREIGN KEY (player) REFERENCES players(id)
 );
